refactor referrals into smalled components
