+++ src/views/Referrals/EmptyMessage.js
@@ -1,0 +1,15 @@
+import Tooltip from "../../components/Tooltip/Tooltip";
+
+function EmptyMessage({ message = "", tooltipText }) {
+  return (
+    <div className="empty-message">
+      {tooltipText ? (
+        <Tooltip handle={<p>{message}</p>} position="center-bottom" renderContent={() => tooltipText} />
+      ) : (
+        <p>{message}</p>
+      )}
+    </div>
+  );
+}
+
+export default EmptyMessage;
